app/requirements.txt
#sqlalchemy==2.0.28
azure-identity==1.15.0
pyodbc==5.1.0
Flask-SQLAlchemy==3.1.1